[rename]파일 위치 이동 (#42)
`SimpleBrandDto`의 위치를 변경
