+++ src/main/java/org/kakaoshare/backend/domain/brand/repository/query/BrandRepositoryCustom.java
@@ -1,6 +1,6 @@
 package org.kakaoshare.backend.domain.brand.repository.query;
 
-import org.kakaoshare.backend.domain.brand.entity.query.SimpleBrandDto;
+import org.kakaoshare.backend.domain.brand.dto.SimpleBrandDto;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
